backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pymongo==4.10.1
pydantic==2.12.3
python-dotenv==1.0.1
httpx==0.28.1
grpcio==1.78.0
grpcio-tools==1.78.0
protobuf==6.31.1
